Fix memory leak
The Play component calls loadGame, and if Play is remounted before loadGame
finishes, loadGame will try to update state on the old unmounted Play
component.  This is a memory leak, so to fix this, we ensure that Play is only
mounted once.

Play was remounted whenever the Routing component was rerendered (i.e. once
the user loaded).  To prevent this, we set the render parameter on the
Route component.  We do the same for all of the other routes that take
parameters.